# Conflicts:
#	composer.json
#	composer.lock
#	routes/web.php

+++ routes/api.php
@@ -1,8 +1,9 @@
 <?php
 
+use App\Http\Api\AuthController;
 use App\Http\Api\ProductsController;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Http\Request;
 
 /*
 |--------------------------------------------------------------------------
@@ -15,12 +16,18 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
+Route::middleware('auth:sanctum','verified')->get('/user', function (Request $request) {
     return $request->user();
 });
 
+Route::post('register', [AuthController::class, 'register']);
+Route::post('login', [AuthController::class, 'login']);
+Route::post('me', [AuthController::class, 'me'])->middleware('auth:sanctum');
+
+
+
 Route::get('products', [ProductsController::class, 'index']);
 Route::get('product/{id}', [ProductsController::class, 'show']);
-Route::put('product/{id}', [ProductsController::class, 'update']);
-Route::delete('product/delete/{id}', [ProductsController::class, 'delete']);
-Route::post('product', [ProductsController::class, 'add']);
+Route::put('product/{id}', [ProductsController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('product/delete/{id}', [ProductsController::class, 'delete'])->middleware('auth:sanctum');
+Route::post('product', [ProductsController::class, 'add'])->middleware('auth:sanctum');
